Macros: Require only "stable" version of selinux-policy
In special circumstances it is possible that selinux-policy used to
build a DSP package is later not available for installation.

Work around this problem by only recommending the latest policy version
and adding a fallback "Requires" to a hardcoded "stable" version. This
version should be updated when major policy changes take place.

Signed-off-by: Vit Mojzis <vmojzis@redhat.com>